pyproject.toml
[tool.poetry]
name = "data-agent"
version = "0.1.0"
description = ""
authors = ["ChihYu Yeh <chihyu.jimmy.yeh@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
smolagents = "^1.10.0"
datasets = "^3.3.2"
python-dotenv = "^1.0.1"
dabstep-benchmark = {git = "https://git@hf.co/spaces/adyen/DABstep.git", rev = "main"}
ipykernel = "^6.29.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
